Implemented Distributional Matching database

+++ src/edu/macalester/moviedatabase/TagSimilarityMeasure.java
@@ -1,5 +1,5 @@
 package edu.macalester.moviedatabase;
 
 public interface TagSimilarityMeasure {
-	int calculateSimilarity(String tag1, String tag2);
+	double calculateSimilarity(String tag1, String tag2);
 }
